refactor: isr movie detail images
