+++ src/components/pages/movie/images/types.ts
@@ -1,0 +1,8 @@
+import { MovieDetailPageParams } from "components/pages/movie/detail/types";
+import { MovieImagesResponse } from "services/tmdb/movie/images/types";
+
+export type MovieImagesPageProps = {
+  fallbackData?: MovieImagesResponse;
+};
+
+export type MovieImagesPageParams = MovieDetailPageParams;
