+++ src/pages/index.tsx
@@ -1,23 +1,2 @@
-import { GetStaticProps } from "next";
-
-import { HomePageProps } from "components/pages/home/types";
-import { getMovieListServer } from "services/tmdb/movie/list";
-
-export const getStaticProps: GetStaticProps<HomePageProps> = async () => {
-  const popularFallbackData = await getMovieListServer("popular");
-  const nowPlayingFallbackData = await getMovieListServer("now_playing");
-  const topRatedFallbackData = await getMovieListServer("top_rated");
-  const upcomingFallbackData = await getMovieListServer("upcoming");
-
-  return {
-    props: {
-      popularFallbackData,
-      nowPlayingFallbackData,
-      topRatedFallbackData,
-      upcomingFallbackData,
-    },
-    revalidate: 60,
-  };
-};
-
+export { getStaticProps } from "components/pages/home/loader";
 export { default } from "components/pages/home";
