+++ src/pages/movie/[id]/images.tsx
@@ -1,1 +1,5 @@
-export { default } from "components/pages/movie/detail/images";
+export {
+  getStaticPaths,
+  getStaticProps,
+} from "components/pages/movie/images/loader";
+export { default } from "components/pages/movie/images";
